Movement script fully separated
More things to extract from the old player movement still
