+++ Assets/Player/Scripts/Move.cs
@@ -4,317 +4,111 @@ using UnityEngine;
 
 public class Move : MonoBehaviour
 {
-    [Header("Keybinds")]
-    public KeyCode jumpKey = KeyCode.Space;
-    public KeyCode sprintKey = KeyCode.LeftShift;
-    public KeyCode crouchKey = KeyCode.LeftControl;
-
-    [Header("References")]
-    public ParticleSystem ps;
-
-    [Header("Movement")]
-    public float walkSpeed;
-    public float sprintSpeed;
-    public float slideSpeed;
-    public float wallRunSpeed;
-
-    [HideInInspector]
-    public float moveSpeed;
-
-    private float desiredMoveSpeed;
-    private float lastDesiredMoveSpeed;
-
-    public float speedIncreaseMultiplier;
-    public float slopeIncreaseMultiplier;
-
-    public float strafeDampener;
-
-    [Header("Drag")]
-    public float groundDrag;
-
-    public float airMultiplier;
-    [Range(0, 1)]
-    public float airDecay;
-    private float airDrag = 1;
-    private float startAirDrag;
-    
-    [Header("Crouching")]
-    public float crouchSpeed;
-    public float crouchYScale;
-    private float startYScale;
-
-    [Header("Ground Check")]
-    public float playerHeight;
-    public LayerMask whatIsGround;
-    
-
-    [Header("Slope Handling")]
-    public float maxSlopeAngle;
-    public float slopeMultiplier;
-    private RaycastHit slopeHit;
-    private bool exitingSlope;
+    private PlayerMovementStateMachine pm;
+    private Rigidbody rb;
 
     public Transform orientation;
 
-    float horizontalInput;
-    float verticalInput;
+    [Header("Speeds")]
+    public float forwardSpeed;
+    public float strafeSpeed;
+    public float backwardSpeed;
+    [Range(0, 1)] public float airMultipler;
 
-    Vector3 moveDirection;
+    [Header("Drags")]
+    [Range(0, 1)] public float groundDrag;
+    [Range(0, 1)] public float airDrag;
 
-    [HideInInspector]
-    public Rigidbody rb;
+    [Header("Stick to Ground")]
+    public bool stickToGround;
+    public float stickingForce;
 
-    [Header("Flags")]
-    public bool sliding;
-    public bool grounded;
-    public bool wallRunning;
-
-    public MovementState state;
-
-    public enum MovementState
+    // Start is called before the first frame update
+    void Start()
     {
-        walking,
-        sprinting,
-        wallRunning,
-        crouching,
-        sliding,
-        air
-    }
-
-    private void Start()
-    {
+        pm = GetComponent<PlayerMovementStateMachine>();
         rb = GetComponent<Rigidbody>();
-        rb.freezeRotation = true;
-
-        startYScale = transform.localScale.y;
-
-        startAirDrag = airDrag;
-    }
-
-    private void Update()
-    {
-        //Ground Check
-        grounded = Physics.Raycast(transform.position, Vector3.down, playerHeight * 0.5f + 0.2f, whatIsGround);
-
-        
-        MyInput();
-        SpeedControl();
-        StateHandler();
-
-        //Handle Drag
-        if (grounded)
-            rb.drag = groundDrag;
-        else
-            rb.drag = 0;
-
-        //Speed FX
-        if (rb.velocity.magnitude >= 12)
-        {
-            ps.Play();
-        }
-        else
-        {
-            ps.Stop();
-        }
     }
 
     private void FixedUpdate()
     {
         MovePlayer();
-    }
-
-    private void MyInput()
-    {
-        horizontalInput = Input.GetAxisRaw("Horizontal");
-        verticalInput = Input.GetAxisRaw("Vertical");
-
-        //CROUCH
-        if(Input.GetKeyDown(crouchKey))
-        {
-            transform.localScale = new Vector3(transform.localScale.x, crouchYScale, transform.localScale.z);
-            rb.AddForce(Vector3.down * 5f, ForceMode.Impulse); //Makes Sure Player isn't hovering after crouching
-        }
-
-        //Release Crouch
-        if(Input.GetKeyUp(crouchKey))
-        {
-            transform.localScale = new Vector3(transform.localScale.x, startYScale, transform.localScale.z);
-        }
-    }
-
-    private void StateHandler()
-    {
-        //Mode - WallRunning
-        if (wallRunning)
-        {
-            state = MovementState.wallRunning;
-            desiredMoveSpeed = wallRunSpeed;
-        }
-
-        //Mode - Sliding
-        else if (sliding)
-        {
-            state = MovementState.sliding;
-            if (OnSlope() && rb.velocity.y < 0.1f)
-                desiredMoveSpeed = slideSpeed;
-            else
-                desiredMoveSpeed = sprintSpeed;
-        }
-
-        //Mode - Crouching
-        else if (Input.GetKey(crouchKey))
-        {
-            state = MovementState.crouching;
-            desiredMoveSpeed = crouchSpeed;
-        }
-
-        //Mode - Sprinting
-        else if (grounded && Input.GetKey(sprintKey))
-        {
-            state = MovementState.sprinting;
-            desiredMoveSpeed = sprintSpeed;
-        }
-
-        //Mode - Walking
-        else if (grounded)
-        {
-            state = MovementState.walking;
-            desiredMoveSpeed = walkSpeed;
-        }
-
-        //Mode - Air
-        else
-        {
-            state = MovementState.air;
-        }
-
-        //Check If Desried Speed has Changed
-        if (Mathf.Abs(desiredMoveSpeed - lastDesiredMoveSpeed) > 4f && moveSpeed != 0)
-        {
-            StopAllCoroutines();
-            StartCoroutine(LerpMoveSpeed());
-        }
-        else
-        {
-            moveSpeed = desiredMoveSpeed;
-        }
-        lastDesiredMoveSpeed = desiredMoveSpeed;
-    }
-    //penis
-    private IEnumerator LerpMoveSpeed() //Coroutine
-    {
-        //Interpolates movementSpeed to the Desired Value
-        float time = 0;
-        float difference = Mathf.Abs(desiredMoveSpeed - moveSpeed);
-        float startValue = moveSpeed;
-
-        while (time < difference)
-        {
-            moveSpeed = Mathf.Lerp(startValue, desiredMoveSpeed, time / difference);
-
-            if (OnSlope())
-            {
-                float slopeAngle = Vector3.Angle(Vector3.up, slopeHit.normal);
-                float slopeAngleIncrease = 1 + (slopeAngle / 90f);
-
-                time += Time.deltaTime * speedIncreaseMultiplier * slopeIncreaseMultiplier * slopeAngleIncrease;
-            }
-            else
-                time += Time.deltaTime * speedIncreaseMultiplier;
-
-            yield return null;
-        }
-
-        moveSpeed = desiredMoveSpeed;
+        ApplyDrag();
     }
 
     private void MovePlayer()
     {
-        //Calculate Movement Dir
-        moveDirection = orientation.forward * verticalInput + orientation.right * horizontalInput;
-
-        //Dampen Strafe Speed
-        if (horizontalInput != 0 && verticalInput == 0)
-        {
-            moveSpeed *= strafeDampener;
-        }
-
-        //On Slope
-        if (OnSlope() && !exitingSlope)
-        {
-            rb.AddForce(GetSlopeMoveDirection(moveDirection) * moveSpeed * slopeMultiplier * 10f, ForceMode.Force);
-
-            if (rb.velocity.y > 0)
-            {
-                rb.AddForce(Vector3.down * 80f, ForceMode.Force);
-            }
-        }
-
-        //On Ground
-        else if (grounded)
-        {
-            rb.AddForce(moveDirection.normalized * moveSpeed * 10f, ForceMode.Force);
-            airDrag = startAirDrag;
-        }
-
-        //In Air
-        
-        else if (!grounded)
-        {
-            //Handle Air Drag If No Forward Input
-            if (verticalInput == 0)
-            {
-                airDrag -= (airDecay * 0.01f);
-                rb.velocity = new Vector3(rb.velocity.x *  airDrag, rb.velocity.y, rb.velocity.z * airDrag);
-            }
-
-            rb.velocity = rb.velocity * (1-startAirDrag*Time.deltaTime);
-
-            rb.AddForce(moveDirection.normalized * moveSpeed * airMultiplier * 10f, ForceMode.Force);
-        }
-    
-        //Disable Gravity On Slope (Prevents Sliding)
-        if(!wallRunning) rb.useGravity = !OnSlope();
+        if (pm.state == PlayerMovementStateMachine.MovementState.normal)
+            MovePlayer(1);
+        if (pm.state == PlayerMovementStateMachine.MovementState.falling)
+            MovePlayer(airMultipler);
+        if (pm.state == PlayerMovementStateMachine.MovementState.jumping)
+            MovePlayer(airMultipler);
     }
 
-    private void SpeedControl()
+    private void MovePlayer(float moveMultiplier)
     {
-        
-        //Limit Speed on Slope
-        if (OnSlope() && !exitingSlope)
-        {
-            if(rb.velocity.magnitude > moveSpeed)
-                rb.velocity = rb.velocity.normalized * moveSpeed;
-        }
+        //strafe
+        Vector3 moveVector = orientation.right * pm.horizontalInput * strafeSpeed;
 
-        //Limit Speed on Ground and Air
+        //add front-back
+        if (pm.verticalInput > 0)
+            moveVector += orientation.forward * pm.verticalInput * forwardSpeed;
         else
-        {
-            Vector3 flatVelocity = new Vector3(rb.velocity.x, 0f, rb.velocity.z);
+            moveVector += orientation.forward * pm.verticalInput * backwardSpeed;
 
-            //Limit Velocity
-            if (flatVelocity.magnitude > moveSpeed)
-            {
-                Vector3 limitedVelocity = flatVelocity.normalized * moveSpeed;
-                rb.velocity = new Vector3(limitedVelocity.x, rb.velocity.y, limitedVelocity.z);
-            }
+        //apply ground adjustments
+        moveVector = AdjustIfGrounded(moveVector);
 
-        }
+        //apply multiplier (reduces movement forces while in air)
+        moveVector *= moveMultiplier;
+
+        //apply movement force
+        rb.AddForce(moveVector, ForceMode.Impulse);
+
     }
 
-    public bool OnSlope()
+    private Vector3 AdjustIfGrounded(Vector3 moveVector)
     {
-        if(Physics.Raycast(transform.position, Vector3.down, out slopeHit, playerHeight * 0.5f + 0.3f))
-        {
-            float angle = Vector3.Angle(Vector3.up, slopeHit.normal);
-            return angle < maxSlopeAngle && angle != 0;
-        }
+        //if not grounded
+        if (pm.state != PlayerMovementStateMachine.MovementState.normal)
+            return moveVector;
 
-        return false;
+        //perform raycast
+        RaycastHit hit;
+        Physics.Raycast(transform.position, Vector3.down, out hit, pm.playerHeight * 0.5f + 0.3f);
+
+        //find upward normal
+        Vector3 upwardNormal = hit.normal;
+        if (Vector3.Dot(hit.normal, Vector3.up) < 0)
+            upwardNormal *= -1;
+
+        //adjust move direction for slope
+        moveVector = Vector3.ProjectOnPlane(moveVector, upwardNormal);
+
+        //TODO re-semi-normalize
+
+        //apply ground stickiness
+        if (stickToGround)
+            rb.AddForce(-upwardNormal * stickingForce, ForceMode.Impulse);
+
+        return moveVector;
     }
 
-    public Vector3 GetSlopeMoveDirection(Vector3 direction)
+    private void ApplyDrag()
     {
-        return Vector3.ProjectOnPlane(direction, slopeHit.normal).normalized;
+        if (pm.state == PlayerMovementStateMachine.MovementState.normal)
+            ApplyDrag(groundDrag);
+        if (pm.state == PlayerMovementStateMachine.MovementState.falling)
+            ApplyDrag(airDrag);
+        if (pm.state == PlayerMovementStateMachine.MovementState.jumping)
+            ApplyDrag(airDrag);
     }
+
+    private void ApplyDrag(float drag)
+    {
+        Vector3 moveVel = new Vector3(rb.velocity.x, 0, rb.velocity.z);
+        rb.velocity -= moveVel * drag;
+    }
+
 }
