+++ Assets/Player/Scripts/Normal.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f91c9ad812ddc459c98145121f698187
+guid: f4587ddc1e221420e97f355c0fda412d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
