+++ Assets/World/Frictionless.physicMaterial
@@ -8,7 +8,7 @@ PhysicMaterial:
   m_PrefabAsset: {fileID: 0}
   m_Name: Frictionless
   dynamicFriction: 0
-  staticFriction: 0
+  staticFriction: 0.9
   bounciness: 0
   frictionCombine: 2
   bounceCombine: 2
